+++ src/problem/non_abundant_sums.rs
@@ -2,13 +2,13 @@ use std::collections::BTreeSet;
 use crate::{number_theory::fast_pow, primes::prime_factors};
 use super::Problem;
 
-const MAX_NON_ABUNDANT: usize = 28123;
+const MAX_NON_ABUNDANT: u64 = 28123;
 
 // This constant is based on the wikipedia article: https://en.wikipedia.org/wiki/Abundant_number
 // all numbers greater than this CAN be written as the sum of two abundant numbers
-const MAX_NON_ABUNDANT_SUM: usize = 20161;
+const MAX_NON_ABUNDANT_SUM: u64 = 20161;
 
-fn sigma(n: usize) -> usize {
+fn sigma(n: u64) -> u64 {
     let prime_factors = prime_factors(n);
     let mut divisors_sum = 1;
 
@@ -19,7 +19,7 @@ fn sigma(n: usize) -> usize {
     divisors_sum
 }
 
-fn is_abundant_sum(n: usize, lookup: &BTreeSet<usize>) -> bool {
+fn is_abundant_sum(n: u64, lookup: &BTreeSet<u64>) -> bool {
     if n > MAX_NON_ABUNDANT {
         return true;
     }
@@ -52,7 +52,7 @@ impl Problem for NonAbundantSumsProblem {
     }
 
     fn solve(&self) -> String {
-        let mut abundant_numbers: BTreeSet<usize> = BTreeSet::new();
+        let mut abundant_numbers: BTreeSet<u64> = BTreeSet::new();
 
         let mut n = 1;
         while n <= MAX_NON_ABUNDANT {
